+++ src/Stack/Setup.hs
@@ -1051,9 +1051,9 @@ warnUnsupportedCompiler ghcVersion = do
           , flow "Valiantly attempting to run anyway, but this is doomed."
           ]
         pure True
-    | ghcVersion >= mkVersion [9, 13] && notifyIfGhcUntested -> do
+    | ghcVersion >= mkVersion [9, 15] && notifyIfGhcUntested -> do
         prettyWarnL
-          [ flow "Stack has not been tested with GHC versions 9.14 and above, \
+          [ flow "Stack has not been tested with GHC versions 9.16 and above, \
                  \and using"
           , fromString (versionString ghcVersion) <> ","
           , flow "this may fail."
@@ -1095,9 +1095,9 @@ warnUnsupportedCompilerCabal cp didWarn = do
           , parens (style Shell "nightly-2018-03-13")
           , flow "or later specify such GHC versions."
           ]
-    | cabalVersion >= mkVersion [3, 15] && notifyIfCabalUntested ->
+    | cabalVersion >= mkVersion [3, 17] && notifyIfCabalUntested ->
         prettyWarnL
-          [ flow "Stack has not been tested with Cabal versions 3.16 and \
+          [ flow "Stack has not been tested with Cabal versions 3.18 and \
                  \above, but version"
           , fromString (versionString cabalVersion)
           , flow "was found, this may fail."